+++ packages/moonstone/Scroller/ScrollableNative.js
@@ -954,7 +954,6 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 		updateScrollThumbSize = () => {
 			const
 				{horizontalScrollbar, verticalScrollbar} = this.props,
-				{isHorizontalScrollbarVisible, isVerticalScrollbarVisible} = this.state,
 				bounds = this.getScrollBounds(),
 				canScrollHorizontally = this.canScrollHorizontally(bounds),
 				canScrollVertically = this.canScrollVertically(bounds),
@@ -969,8 +968,7 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 						...bounds,
 						scrollLeft: this.scrollLeft,
 						scrollTop: this.scrollTop
-					},
-					spotItem = Spotlight.getCurrent();
+					};
 
 				if (curHorizontalScrollbarVisible) {
 					this.horizontalScrollbarRef.update(updatedBounds);