+++ packages/moonstone/CHANGELOG.md
@@ -6,6 +6,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Fixed
 
+- `moonstone/Scrollable` to not show scroll thumb when only child item is updated
 - `moonstone/SliderDecorator` to not unnecessarily fire `onChange` if the initial value has not changed
 
 ## [1.13.0] - 2017-11-28